Improve submitter import performance

+++ Importer/PKP/ArticleSubmitterImporter.php
@@ -2,6 +2,7 @@
 
 namespace Ojs\ImportBundle\Importer\PKP;
 
+use Ojs\ImportBundle\Entity\PendingSubmitterImport;
 use Ojs\JournalBundle\Entity\Article;
 use Ojs\ImportBundle\Importer\Importer;
 
@@ -12,22 +13,41 @@ class ArticleSubmitterImporter extends Importer
      */
     public function importArticleSubmitter($userImporter)
     {
-        $pendingImports = $this->em->getRepository('ImportBundle:PendingSubmitterImport')->findAll();
         $this->consoleOutput->writeln("Importing article submitters...");
 
-        foreach ($pendingImports as $import) {
+        $result = $this->em
+            ->getRepository(PendingSubmitterImport::class)
+            ->createQueryBuilder('import')
+            ->select('import.id')
+            ->getQuery()->getScalarResult();
+        $ids = array_column($result, 'id');
+        $counter = 0;
+
+        foreach ($ids as $id) {
+            $import = $this->em->find(PendingSubmitterImport::class, $id);
+
+            if (!$import || !$import->getArticle()) {
+                continue;
+            }
+
             $user = $userImporter->importUser($import->getOldId());
 
             if ($user) {
                 /** @var Article $article */
                 $article = $import->getArticle();
                 $article->setSubmitterUser($user);
-
                 $this->em->persist($article);
             }
 
             $this->em->remove($import);
-            $this->em->flush();
+            $counter++;
+
+            if ($counter % 10 == 0) {
+                $this->em->flush();
+                $this->em->clear();
+            }
         }
+
+        $this->em->flush();
     }
 }
